+++ src/main/java/com/dreamypatisiel/devdevdev/domain/repository/MemberRepository.java
@@ -2,6 +2,8 @@ package com.dreamypatisiel.devdevdev.domain.repository;
 
 import com.dreamypatisiel.devdevdev.domain.entity.Member;
 import com.dreamypatisiel.devdevdev.domain.entity.SocialType;
+import com.dreamypatisiel.devdevdev.domain.entity.embedded.Email;
+import jakarta.persistence.Embeddable;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,6 +11,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface MemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findMemberByUserIdAndSocialType(String userId, SocialType socialType);
     List<Member> findMembersByUserIdAndSocialType(String userId, SocialType socialType);
-
-    Optional<Member> findMemberByEmail(String email);
+    Optional<Member> findMemberByEmailAndSocialType(Email email, SocialType socialType);
+    Optional<Member> findMemberByRefreshToken(String refreshToken);
 }
