--- src/main/java/com/dreamypatisiel/devdevdev/global/config/jwt/model/Token.java
@@ -1,12 +1,0 @@
-package com.dreamypatisiel.devdevdev.global.config.jwt.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class Token {
-    private String accessToken;
-    private String refreshToken;
-}
