--- src/main/java/com/dreamypatisiel/devdevdev/global/config/security/oauth2/OAuth2Config.java
@@ -1,14 +1,0 @@
-package com.dreamypatisiel.devdevdev.global.config.security.oauth2;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
-
-@Configuration
-public class OAuth2Config {
-
-    @Bean
-    public DefaultOAuth2UserService defaultOAuth2UserService() {
-        return new DefaultOAuth2UserService();
-    }
-}
