+++ src/main/java/com/dreamypatisiel/devdevdev/global/security/SecurityConstant.java
@@ -1,0 +1,7 @@
+package com.dreamypatisiel.devdevdev.global.security;
+
+public class SecurityConstant {
+    public static final String SPRING_H2_CONSOLE_ENABLED = "spring.h2.console.enabled";
+    public static final String WILDCARD_PATTERN = "/**";
+    public static final long PREFLIGHT_MAX_AGE = 3600L;
+}
