+++ src/main/java/com/dreamypatisiel/devdevdev/global/security/jwt/CookieUtils.java
@@ -1,0 +1,75 @@
+package com.dreamypatisiel.devdevdev.global.security.jwt;
+
+import com.dreamypatisiel.devdevdev.global.security.jwt.model.JwtCookieConstant;
+import com.dreamypatisiel.devdevdev.global.security.jwt.model.Token;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.util.SerializationUtils;
+
+import java.util.Arrays;
+import java.util.Base64;
+import java.util.Optional;
+
+public class CookieUtils {
+
+    public static final int DEFAULT_MAX_AGE = 180;
+
+    public static Optional<Cookie> getCookie(HttpServletRequest request, String name) {
+        Cookie[] cookies = request.getCookies();
+        if (cookies != null && cookies.length > 0) {
+            for (Cookie cookie : cookies) {
+                if (cookie.getName().equals(name)) {
+                    return Optional.of(cookie);
+                }
+            }
+        }
+        return Optional.empty();
+    }
+
+    public static Optional<String> readServletCookie(HttpServletRequest request, String name) {
+        return Arrays.stream(request.getCookies())
+                .filter(cookie -> name.equals(cookie.getName()))
+                .map(Cookie::getValue)
+                .findAny();
+    }
+
+    public static void addCookie(HttpServletResponse response, String name, String value, int maxAge, boolean isHttpOnly, boolean isSecure) {
+        Cookie cookie = new Cookie(name, value);
+        cookie.setPath("/");
+        cookie.setHttpOnly(isHttpOnly);
+        cookie.setSecure(isSecure);
+        cookie.setMaxAge(maxAge);
+        response.addCookie(cookie);
+    }
+
+    public static void deleteCookie(HttpServletRequest request, HttpServletResponse response, String name) {
+        Cookie[] cookies = request.getCookies();
+        if (cookies != null && cookies.length > 0) {
+            for (Cookie cookie : cookies) {
+                if (cookie.getName().equals(name)) {
+                    cookie.setValue("");
+                    cookie.setPath("/");
+                    cookie.setMaxAge(0);
+                    response.addCookie(cookie);
+                }
+            }
+        }
+    }
+
+    public static String serialize(Object object) {
+        return Base64.getUrlEncoder()
+                .encodeToString(SerializationUtils.serialize(object));
+    }
+
+    public static <T> T deserialize(Cookie cookie, Class<T> cls) {
+        return cls.cast(SerializationUtils.deserialize(Base64.getUrlDecoder().decode(cookie.getValue())));
+    }
+
+    public static void configJwtCookie(HttpServletResponse response, Token token) {
+        CookieUtils.addCookie(response, JwtCookieConstant.DEVDEVDEV_ACCESS_TOKEN,
+                token.getAccessToken(), CookieUtils.DEFAULT_MAX_AGE, false, false);
+        CookieUtils.addCookie(response, JwtCookieConstant.DEVDEVDEV_REFRESH_TOKEN,
+                token.getRefreshToken(), CookieUtils.DEFAULT_MAX_AGE, true, true);
+    }
+}
