+++ src/main/java/com/dreamypatisiel/devdevdev/global/security/jwt/model/JwtClaimConstant.java
@@ -1,0 +1,6 @@
+package com.dreamypatisiel.devdevdev.global.security.jwt.model;
+
+public class JwtClaimConstant {
+    public static String email = "email";
+    public static String socialType = "socialType";
+}
