+++ src/main/java/com/dreamypatisiel/devdevdev/global/security/jwt/model/JwtCookieConstant.java
@@ -1,0 +1,6 @@
+package com.dreamypatisiel.devdevdev.global.security.jwt.model;
+
+public class JwtCookieConstant {
+    public static final String DEVDEVDEV_REFRESH_TOKEN = "DEVDEVDEV_REFRESH_TOKEN";
+    public static final String DEVDEVDEV_ACCESS_TOKEN = "DEVDEVDEV_ACCESS_TOKEN";
+}
