+++ src/main/java/com/dreamypatisiel/devdevdev/global/security/jwt/model/Token.java
@@ -1,0 +1,12 @@
+package com.dreamypatisiel.devdevdev.global.security.jwt.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class Token {
+    private String accessToken;
+    private String refreshToken;
+}
