+++ src/main/java/com/dreamypatisiel/devdevdev/global/security/jwt/service/JwtMemberService.java
@@ -1,0 +1,24 @@
+package com.dreamypatisiel.devdevdev.global.security.jwt.service;
+
+import com.dreamypatisiel.devdevdev.domain.entity.Member;
+import com.dreamypatisiel.devdevdev.domain.repository.MemberRepository;
+import com.dreamypatisiel.devdevdev.domain.service.response.MemberResponse;
+import com.dreamypatisiel.devdevdev.exception.MemberException;
+import java.util.Optional;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class JwtMemberService {
+
+    private final MemberRepository memberRepository;
+
+    public MemberResponse registerRefreshToken(String refreshToken) {
+        Member findMember = memberRepository.findMemberByRefreshToken(refreshToken)
+                .orElseThrow(() -> new MemberException(MemberException.INVALID_MEMBER_NOT_FOUND_MESSAGE));
+
+        findMember.updateRefreshToken(refreshToken);
+        return null;
+    }
+}
