+++ src/main/java/com/dreamypatisiel/devdevdev/web/response/ResultType.java
@@ -1,0 +1,6 @@
+package com.dreamypatisiel.devdevdev.web.response;
+
+public enum ResultType {
+    SUCCESS,
+    FAIL
+}
